requirements.txt
argparse
pandas
pillow
tqdm
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
peft>=0.5.0
accelerate>=0.20.0